pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "T2IScoreScore"
version = "0.1"
description = "A framework for evaluating text-to-image metrics using T2IScoreScore (TS2)"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "tqdm>=4.62.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "pillow>=9.0.0",
    "openai>=1.0.0",
    "sentence-transformers>=2.2.0",
    "bitsandbytes>=0.41.0",
    "accelerate>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=22.0",
    "isort>=5.0",
    "pytest-cov>=2.12.0",
    "mypy>=0.910",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["T2IMetrics*", "T2IScoreScore*"]

[tool.setuptools.package-data]
T2IMetrics = ["resources/**/*"]
T2IScoreScore = ["resources/**/*"] 